Update acnagor.msg
If playing a female character than in German, it must be Cousine instead of Cousin. 

Since I don't know if switching here to female would have done the trick, too.:
https://tra.bgforge.net/translate/fallout/rp/de/?checksum=&offset=&q=Cousin%2C+am+I+glad+to+see+you%21+I+need+your+help.&search=ftx&source=on&target=on&type=all&ignored=False&date=&only_user=&exclude_user=

That's why I decided to download the acnagor.msg and edit it manually. I placed it in the according folders at the fallout 2 and tried it. It worked fine, now when playing as female it's correctly displayed. I also fixed some minor grammar errors.